% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dst}
\alias{dst}
\title{Example distance matrix}
\format{A 613 x 613 matrix with column and row names corresponding to VET images}
\usage{
dst
}
\description{
A distance matrix of all VET images created with VGG-16
}
\keyword{datasets}
