% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_corr_matrix.R
\name{get_corr_matrix}
\alias{get_corr_matrix}
\title{Correlation matrix from response probabilities}
\usage{
get_corr_matrix(response_lists)
}
\arguments{
\item{response_lists}{The output from \code{\link{run_vet}}}
}
\value{
A correlation matrix
}
\description{
\code{get_cor_mat} gets the correlation matrix where each element corresponds
to the correlation between two category's predicted accuracy
}
