% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{make_parameter_list}
\alias{make_parameter_list}
\title{Makes a list of parameters}
\usage{
make_parameter_list(num_subjects, par_names, lower, upper)
}
\arguments{
\item{num_subjects}{The number of subject parameters}

\item{par_names}{A character vector of parameter names}

\item{lower}{The lower bound on the uniform distribution}

\item{upper}{The upper bound on the uniform distribution}
}
\description{
\code{make_parameter_list} makes a list of parameters for input to \code{\link{run_vet}}
}
\examples{
#make parameters for 10 subjects
par_list = make_parameter_list(num_subjects=10,par_names=c('c','beta'),lower=c(0,0),upper=c(1,1))
}
