% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_vet_list.R
\name{make_vet_list}
\alias{make_vet_list}
\title{Prepare the VET test list}
\usage{
make_vet_list(vet_test_file)
}
\arguments{
\item{vet_test_file}{the VET 1.0 test list}
}
\value{
The processed VET that can be used in \code{run_vet}
}
\description{
\code{make_vet_list} Takes the raw VET and prepares it for use in \code{\link{run_vet}}
}
\examples{
vet_list = make_vet_list(system.file('extdata','vet_1.0_test_list.csv',package='vetnet'))
}
