% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_vet.R
\name{run_vet}
\alias{run_vet}
\title{Simulate the Vanderbilt Expertise Test}
\usage{
run_vet(vet, dst, model, par)
}
\arguments{
\item{vet}{The preprocessed VET test list made with \code{\link{make_vet_list}}}

\item{dst}{A distance matrix or list of distance matrices. Each element (i,j) is the distance between image i and j.
The name of row i and column j correspond to the name of image i and name of image j, respectively.}

\item{model}{A function that returns the probability of choosing the target probe.
Its first argument is a 3 element vector containing the summed distances for a given probe and all study items.
Its second argument is a named vector of parameters used to convert the distance into a probabilty of picking the target.
See the example below for further explanation.}

\item{par}{A data frame of parameter values}
}
\value{
A data frame where row i correspond to the predicted probability of choosing the target on trial i.
}
\description{
\code{run_vet} gets the probability of choosing the target image for all trials on the VET
}
\examples{
#make a list of parameters
par_list = make_parameter_list(num_subjects=10,par_names=c('c','beta'),
                               lower=c(0,0),upper=c(1,1))

# set the model function
model = function(d,pars) {
    c = pars$c
    beta = pars$beta
    p = sapply(d, function(d) exp(-c*d) / (exp(-c*d) + beta))
    p = p[1] / sum(p) #probability of picking target
}

#run the simulation
response_lists = run_vet(vet_list, dst, model, par_list)

#get the correlation matrix
get_corr_matrix(response_lists)
}
