% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vet_list}
\alias{vet_list}
\title{The preprocessed VET test list}
\format{A list with two elements: study and probe
\describe{
  \item{study}{study list for each of the 8 categories}
  \item{probe}{a matrix where each row is the probe names for the given trial,
  the first element is the target}
  ...
}}
\usage{
vet_list
}
\description{
The preprocessed VET test list
}
\keyword{datasets}
